fix: update ssl context to prevent deprecation errors
fix: replace pkg_resources with importlib to prevent deprecation errors
refactor: rename earthdata.py to api.py
