+++ earthdata/__init__.py
@@ -1,4 +1,4 @@
-from earthdata.earthdata import earthdata
+from earthdata.api import api
 import earthdata.version
 # get version number
 __version__ = earthdata.version.version
